refactor: move DocReference to datashare-api text package
