+++ datashare-api/src/test/java/org/icij/datashare/text/DocReferenceTest.java
@@ -1,4 +1,4 @@
-package org.icij.datashare.tasks;
+package org.icij.datashare.text;
 
 import org.junit.Test;
